// src/app/shared/model/curr-obj.model.ts
export class CurrObj {
    RefCurrId: any;
    CurrName: any;
    CurrCode: any;
    RoundedAmt: any;
    MinRefundAmt: any;
    ToleranceAmt: any;
    CofId: any;
    RegRptCode: any;
    LbppCode: any;
    IsActive: any;
    RowVersion: any;
  constructor() { this.RefCurrId = 0; this.RowVersion = "" }
}